pyproject.toml
[project]
name = "secbrain-obsidian-exporter"
version = "0.1.0"
description = "Export SecBrain notes to Obsidian vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
export-notes = "exporter:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["exporter"]
